mark row as available updates ui properly
